==> markdown/sections/01_data_processing.Rmd <==
---
title: "01_data_processing"
output: html_document
---

PLISA PAHO portal only permits download of all countries cumulative data sets in a week by week fashion, by moving the epidemiological week slider while selecting all countries. These require the file format to be encoded differently, and re-saved in this format for further processing. 

The files are loaded and a new directory created for processed data. 

```{r Load all the raw datasets, encode, and merge into one, eval=FALSE}
file_list<-list.files("data/raw_files/PAHO"); file_list

old_dir<-paste0("data/raw_files/PAHO")
new_dir<-paste0("data/source_files/PAHO")

#Encoding & save again

for(j in 1:length(file_list)){ 
  
  oldfiledir <- paste0(old_dir, "/", file_list[j]) 
  newfiledir <- paste0(new_dir, "/", file_list[j])
  fileConn <- file(newfiledir)
  
  writeLines(readLines(con <- file(oldfiledir, encoding = "UCS-2LE")), fileConn)
  close(fileConn)
  
}

csv_files <-list.files(new_dir, full.names=T, recursive=F, pattern="\\.csv$") #directory names for loading files

csv_files <- csv_files[-c(1)];csv_files
temp <- lapply(csv_files, read.table, sep="\t", colClasses = rep("character",16), header=TRUE)

merge <- rbindlist(temp)
```


```{r, eval=FALSE}
#Column rename and select
all_dt <- merge %>% 
  select(Country.or.Subregion, Serotype, Year, Confirmed, Deaths, 
         Epidemiological.Week..a., Severe.Dengue..d., Total.of.Dengue.Cases..b.) %>% 
  rename(adm_0_name = Country.or.Subregion,
         year = Year, 
         epidemiological_week = Epidemiological.Week..a.,
         severe_dengue_cumulative = Severe.Dengue..d.,
         dengue_total_cumulative = Total.of.Dengue.Cases..b.)  %>%
  filter(!epidemiological_week %in% c("-")) # * AL: why does it have some data with epidemiological week being "-" here? 

#convert epidemiological week and year to numeric
all_dt$epidemiological_week <- as.numeric(all_dt$epidemiological_week)
all_dt$year <- as.numeric(all_dt$year)

#remove duplicates
all_dt <- all_dt %>% 
  distinct()%>%
  arrange(adm_0_name, year, epidemiological_week)

#add the end date of the cumulative count for each row ("cum_end_date"). 
for(i in 1:nrow(all_dt)){
 all_dt$cum_end_date[i] <- as.character(epiweekToDate(all_dt$year[i], all_dt$epidemiological_week[i])$d1)

} #*AL: maybe try "lubridate::epiweek()" here? it's faster. 

#add cum_start_date based on the cum_end_date of the previous row. This is done by lagging cum_end_week because there may be cases where the cumulative count is not reported for consecutive weeks.
all_dt <- all_dt %>%
  arrange(adm_0_name, year, epidemiological_week)%>%
  group_by(adm_0_name, year)%>%
  mutate(cum_start_date = ymd(lag(cum_end_date))+1)

#create "complete" dataset which has a row for every epidemiological week for every year, with NA for the dengue_total count if no data reported for this week. Make sure number of epidemiological weeks in each year is correct according to epidemiological calendars for each year (some have 52 weeks, others have 53).
all_dt <- all_dt %>% ungroup()%>%
  tidyr::complete(adm_0_name, year, epidemiological_week)%>%
  filter(!(year %in% c(2015:2019, 2021:2022) & epidemiological_week == 53))  

all_dt %>%
  group_by(adm_0_name) %>% tally()

#add cum_end_date again for rows that were missed from the original data but were filled by 'complete' function above. Add the start/end date of epidemiological week for all rows("calendar_start_date" and "calendar_end_date"). 

for(i in 1:nrow(all_dt)){
  all_dt$calendar_end_date[i] <- as.character(epiweekToDate(all_dt$year[i], all_dt$epidemiological_week[i])$d1)

  if (is.na(all_dt$cum_end_date[i])==FALSE) next
 all_dt$cum_end_date[i] <- as.character(epiweekToDate(all_dt$year[i], all_dt$epidemiological_week[i])$d1)

}

# add cum_start_date, calendar_start_date and source_id
all_dt <- all_dt %>%
  mutate(cum_end_date = ymd(cum_end_date))%>%
  mutate(cum_start_date = ifelse(is.na(cum_start_date), as.character(ymd(cum_end_date)-6), paste0(cum_start_date)))%>%
  mutate(cum_interval = interval(cum_start_date, cum_end_date) %/% weeks(1)+1)%>%
  mutate(calendar_end_date = ymd(calendar_end_date))%>%
  mutate(calendar_start_date = ymd(calendar_end_date)-6)%>%
  mutate(source_id= ifelse(is.na(dengue_total_cumulative)==FALSE, paste0("all_PAHO_countries_PAHO_2014_to_2022_EW",epidemiological_week), NA))

#summary(all_dt$cum_interval)

```


```{r, eval=FALSE}
#administrative unit codes
all_dt$adm_0_name <- stri_trans_general(all_dt$adm_0_name,"Latin-ASCII")

gaul_codes_world <- read.csv(file="data/ref_data/gaul_codes_world.csv")

gaul_codes_world <- gaul_codes_world %>% 
  select(ADM0_NAME, ADM0_CODE) %>% 
  rename(adm_0_name = ADM0_NAME)%>%
  rename(adm_0_code = ADM0_CODE)%>%
  distinct()

#plyr::count(all_dt$adm_0_name) #52 countries
#plyr::count(gaul_codes_world$adm_0_name) #273 countries

ctry_check <- merge(all_dt, gaul_codes_world, by="adm_0_name", all.x=T)

#check the country names that did not match with gaul names
ctry_check %>% group_by(adm_0_name, adm_0_code)%>% tally() %>% filter(is.na(adm_0_code))

ctry_lookup <- c(
  # gaul name to paho name
"Turks and Caicos islands" = "Turks and Caicos Islands",
"British Virgin Islands" = "Virgin Islands (UK)",
"United States Virgin Islands" = "Virgin Islands (US)"
)
# PAHO countries without gaul codes 
#"Bonaire, Saint Eustatius and Saba" 
#"Curacao"
#"Saint Barthelemy", 
#"Saint Martin", * is this the same country as Sint Maarten?
#"Sint Maarten", 

gaul_codes_world <- gaul_codes_world %>%
  mutate(adm_0_name = recode(adm_0_name, !!!ctry_lookup))

all_dt <- merge(all_dt, gaul_codes_world, by="adm_0_name", all.x=T)
```

Each data set is then processed into two data hierarchies, primary (A) and secondary (B), with calendar dates added. Epidemiological week and year remain, as these are useful for processing later on.  

The case counts are left in cumulative counts at this stage. This may be useful for users who prefer to deal in cumulative counts. here, the calendar start date is fixed at the beginning of epidemiological week one of each year, and the calendar end date is moved forward to account for the cumulative count.   
```{r, eval=FALSE}
#select columns for master table_A and change to date format
all_dt_A <- all_dt %>% 
  select(adm_0_name, adm_0_code, year, epidemiological_week,
         calendar_start_date, calendar_end_date, 
         cum_start_date, cum_end_date, cum_interval,
         dengue_total_cumulative, source_id 
         )%>% 
  mutate(calendar_start_date = ymd(calendar_start_date),
         calendar_end_date = ymd(calendar_end_date), 
         cum_start_date = ymd(cum_start_date), 
         cum_end_date = ymd(cum_end_date))

all_dt_A_tidy <- all_dt_A %>% 
  distinct() %>% 
  mutate_at(c(2), ~replace(., is.na(.), 0)) %>% 
  group_by(adm_0_name, year) %>% 
  arrange(adm_0_name, year, epidemiological_week)

#Check admin codes and fill gaps
admin_NA <- all_dt_A_tidy %>% 
  dplyr::filter(adm_0_code==0)

# *AL: please explain this part - why adm_0_code assigned for Virgin islands only? 
#all_dt_A_tidy$adm_0_code[all_dt_A_tidy$adm_0_name=="Virgin Islands (UK)"]<-"39"
#all_dt_A_tidy$adm_0_code[all_dt_A_tidy$adm_0_name=="Virgin Islands (US)"]<-"258"


write.csv(all_dt_A_tidy,"data/processed_data/PAHO_processed_A.csv", row.names = FALSE)
```


```{r, eval=FALSE}
#Dataset B/Secondary dataset, which contains information on : disease severity, serotype, mortality.

all_dt_B <- all_dt %>% 
  select(adm_0_name, adm_0_code, 
         calendar_start_date, calendar_end_date, 
         dengue_total_cumulative, Deaths, Serotype, severe_dengue_cumulative, source_id)%>% 
  rename(dengue_total=dengue_total_cumulative,
         severe_dengue=severe_dengue_cumulative)%>%
  mutate(calendar_start_date = ymd(calendar_start_date),
         calendar_end_date = ymd(calendar_end_date))%>% 
  filter(dengue_total!= "NA")

all_dt_B_tidy <- all_dt_B %>% 
  distinct() %>% 
  mutate_at(c(2), ~replace(., is.na(.), 0))%>% 
  group_by(adm_0_name) %>% 
  arrange(desc(calendar_start_date))

admin_NA <- all_dt_B_tidy %>% 
  dplyr::filter(adm_0_code==0)

#all_dt_B_tidy$adm_0_code[all_dt_B_tidy$adm_0_name=="Virgin Islands (UK)"]<-"39"
#all_dt_B_tidy$adm_0_code[all_dt_B_tidy$adm_0_name=="Virgin Islands (US)"]<-"258"

all_dt_B_tidy<-all_dt_B_tidy %>% 
  pivot_longer(c(severe_dengue, dengue_total, Deaths),
               names_to="dengue_classification", values_to="total")

all_dt_B_tidy <- all_dt_B_tidy[!(all_dt_B_tidy$dengue_classification == ""), ]
all_dt_B_tidy <- all_dt_B_tidy %>%  
  select(adm_0_name, adm_0_code, calendar_start_date, calendar_end_date, 
         dengue_classification, total, Serotype, source_id)

write.csv(all_dt_B_tidy,"data/processed_data/PAHO_processed_B.csv", row.names = FALSE)
```
